adding base class for command tests and renaming file
