+++ Assets/Mirror/Tests/Editor/CommandTest.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -28,88 +27,8 @@ namespace Mirror.Tests.CommandAttrributeTest
         }
     }
 
-    public class CommandTest
+    public class CommandTest : CommandTestBase
     {
-        private List<GameObject> spawned = new List<GameObject>();
-
-        [SetUp]
-        public void Setup()
-        {
-            Transport.activeTransport = new GameObject().AddComponent<MemoryTransport>();
-
-            // start server/client
-            NetworkServer.Listen(1);
-            NetworkClient.ConnectHost();
-            NetworkServer.SpawnObjects();
-            NetworkServer.ActivateHostScene();
-            NetworkClient.ConnectLocalServer();
-
-            NetworkServer.localConnection.isAuthenticated = true;
-            NetworkClient.connection.isAuthenticated = true;
-
-            ClientScene.Ready(NetworkClient.connection);
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            // stop server/client
-            NetworkClient.DisconnectLocalServer();
-
-            NetworkClient.Disconnect();
-            NetworkClient.Shutdown();
-
-            NetworkServer.Shutdown();
-
-            // destroy left over objects
-            foreach (GameObject item in spawned)
-            {
-                if (item != null)
-                {
-                    GameObject.DestroyImmediate(item);
-                }
-            }
-
-            spawned.Clear();
-
-            NetworkIdentity.spawned.Clear();
-
-            GameObject.DestroyImmediate(Transport.activeTransport.gameObject);
-        }
-
-
-        T CreateHostObject<T>(bool spawnWithAuthority) where T : NetworkBehaviour
-        {
-            GameObject gameObject = new GameObject();
-            spawned.Add(gameObject);
-
-            gameObject.AddComponent<NetworkIdentity>();
-
-            T behaviour = gameObject.AddComponent<T>();
-
-            // spawn outwith authority
-            if (spawnWithAuthority)
-            {
-                NetworkServer.Spawn(gameObject, NetworkServer.localConnection);
-            }
-            else
-            {
-                NetworkServer.Spawn(gameObject);
-            }
-            ProcessMessages();
-
-            Debug.Assert(behaviour.hasAuthority == spawnWithAuthority, $"Behaviour Had Wrong Authority when spawned, This means that the test is broken and will give the wrong results");
-
-            return behaviour;
-        }
-
-        private static void ProcessMessages()
-        {
-            // run update so message are processed
-            NetworkServer.Update();
-            NetworkClient.Update();
-        }
-
         [Test]
         public void CommandIsSentWithAuthority()
         {
