+++ Assets/Mirror/Tests/Editor/CommandTestBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 45bb26f0ce04fb749ac83a28d7590ebc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
